requirement.txt
fastapi
uvicorn[standard]

Jinja2
python-multipart

python-jose[cryptography]
passlib[bcrypt]

pydantic
pydantic-settings
python-dotenv

mysql-connector-python
bcrypt

email-validator
python-dateutil
